+++ blocks/form/block.ts
@@ -1,9 +1,10 @@
 import type { IconName } from '@fortawesome/fontawesome-common-types';
 
 /**
- * An option that is displayed in a dropdown menu.
+ * An option that is displayed in a dropdown menu or radio button field.
  */
 interface Choice {
+  labelText: any;
   /**
    * The label used to display the option.
    */
@@ -12,7 +13,7 @@ interface Choice {
   /**
    * The value to use when selecting the option.
    */
-  value: string;
+  value: any;
 }
 
 interface AbstractField {
@@ -66,6 +67,20 @@ export interface BooleanField extends AbstractField {
    * The type of the field.
    */
   type: 'boolean';
+}
+
+export interface RadioField extends AbstractField {
+  /**
+   * The default value of the field.
+   */
+  defaultValue?: any;
+
+  /**
+   * The list of options the user can select from.
+   */
+  options?: Choice[];
+
+  type: 'radio';
 }
 
 /**
@@ -208,7 +223,8 @@ export type Field =
   | GeoCoordinatesField
   | HiddenField
   | NumberField
-  | StringField;
+  | StringField
+  | RadioField;
 
 export interface InputProps<T, F extends Field> {
   /**