+++ blocks/form/src/components/BooleanInput/index.tsx
@@ -1,4 +1,5 @@
 import { FormattedMessage } from '@appsemble/preact';
+import { FormComponent } from '@appsemble/preact-components/src';
 import classNames from 'classnames';
 import { h, VNode } from 'preact';
 
@@ -17,37 +18,26 @@ export default function BooleanInput({
   value = false,
 }: BooleanInputProps): VNode {
   return (
-    <div className="field is-horizontal">
-      <div className="field-label is-normal">
-        <label className="label" htmlFor={field.name}>
-          {field.label ?? field.name}
-        </label>
-      </div>
-      <div className="field-body">
-        <div className="field">
-          <div className="control">
-            <input
-              checked={!!value}
-              className={classNames('is-checkradio', { 'is-danger': error })}
-              disabled={disabled}
-              id={field.name}
-              name={field.name}
-              onInput={(event) => {
-                onInput(event, (event.target as HTMLInputElement).checked);
-              }}
-              readOnly={field.readOnly}
-              required={field.required}
-              type="checkbox"
-            />
-            <label for={field.name}>{field.labelText ?? field.label ?? null}</label>
-            {error && (
-              <p className={classNames('help', { 'is-danger': error })}>
-                <FormattedMessage id="invalid" />
-              </p>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
+    <FormComponent label={field.label} required={field.required}>
+      <input
+        checked={!!value}
+        className={classNames('is-checkradio', { 'is-danger': error })}
+        disabled={disabled}
+        id={field.name}
+        name={field.name}
+        onInput={(event) => {
+          onInput(event, (event.target as HTMLInputElement).checked);
+        }}
+        readOnly={field.readOnly}
+        required={field.required}
+        type="checkbox"
+      />
+      <label for={field.name}>{field.labelText ?? field.label ?? null}</label>
+      {error && (
+        <p className={classNames('help', { 'is-danger': error })}>
+          <FormattedMessage id="invalid" />
+        </p>
+      )}
+    </FormComponent>
   );
 }